backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-rest-passwordreset
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-allauth
django-import-export
openpyxl